requirements-api.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
celery>=5.3.0
redis>=5.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
requests>=2.31.0
